checker/src/requirements.txt
beautifulsoup4==4.11.1
enochecker3==0.7.0
httpx==0.22.0
python_dateutil==2.8.2
